Point the package dependencies one way and clear the last test debt
Twelve of the thirteen remaining type errors are gone; the one that stays is a
cross-runtime artifact with no honest fix.

Three errors came from `/// <reference lib="deno.ns" />` in shared-test's Deno
entry points, which the npm compiler cannot resolve. @types/deno already
supplies those globals to tsc and Deno supplies them natively, so the
directives are gone and `check:deno` still passes. Three came from
QRtcDataExchanged.description being typed RTCSessionDescription while every
producer feeds it wire JSON and every consumer hands it to setRemoteDescription
-- it is RTCSessionDescriptionInit, the shape without the toJSON a decoded
signal can never carry. Two were real drift in an ungated benchmark: a
syncPublisher option that no longer exists, and a Group literal predating the
formation-lifecycle fields. Two were the mutation-boundary fixtures whose
angle-bracket assertion and parameter property are the syntax under analysis;
nothing imports them, so excluding them from the project keeps the analyzer
reading them as text.

The relic server tests moved to the Deno app they exercise. They imported the
app's source, which imports hono and @api-v1 -- specifiers only Deno resolves --
so the npm project reported them as missing modules. Under `deno test` they
resolve natively. `deno task test` is now wired into test:deno for that app; it
declared the task but only `deno task check` ran, so the tests would otherwise
have moved out of CI entirely. All 36 assertions survive verbatim: std/expect
and std/testing/bdd accept the same matchers, and nothing inspected the spies,
so each vi.fn(impl) became its own implementation.

apps/api-v1 no longer depends on packages/shared-test. The PGlite snapshot
reader needed the app's SQL adapter and the runner's evidence rules, and sat on
the app side reaching back for the second. It now lives beside the rules as an
.mts entry point, which is how shared-test marks Deno-only files so its npm
project skips them, and the dependency points from the runner at the app it
exercises. apps/api-v1/deno.json also drops @shared-web/, which nothing used.

Package tsconfigs now declare only their own dependency closure: shared reaches
nothing, graph and the two runtime halves reach shared, and only shared-test
spans both halves. package-dependency-direction.test.ts enforces the direction
on imports rather than on paths, because a project must still resolve what it
compiles transitively -- shared-test compiles shared-server, so it keeps
@shared-graph without importing it.

The remaining error is Deno.unrefTimer(setInterval(...)) in api-v1: setInterval
returns number under Deno and Timeout under @types/node, and the tests project
loads both. It is correct under the runtime that runs it, and every in-file fix
is a cast.
