+++ apps/api-v1/deno.json
@@ -5,7 +5,6 @@
   "nodeModulesDir": "auto",
   "imports": {
     "@/": "./",
-    "@shared-web/": "../../packages/shared-web/",
     "@shared-server/": "../../packages/shared-server/",
     "@shared-graph/": "../../packages/shared-graph/",
     "@shared-test/": "../../packages/shared-test/",
