+++ package.json
@@ -16,7 +16,7 @@
     "test": "npm run test:unit",
     "test:unit": "vitest run",
     "test:rallar-server-hardening": "vitest run packages/tests/shared-server/queuebox-pubsub-bridge.test.ts packages/tests/shared-server/state-sync-publisher.test.ts packages/tests/shared-server/app-inbox-service.test.ts packages/tests/shared-server/state-sync-cache-hydration.test.ts packages/tests/shared-server/state-event-listing.test.ts packages/tests/api-v1/client-and-group-state-repositories.test.ts && cd apps/api-v1 && deno test --allow-env --allow-read test/services/state-api-resilience-middleware.test.ts test/client-state/client-state-mutation-routes.test.ts test/client-state/client-state-read-routes.test.ts test/group-state/group-admission-routes.test.ts test/group-state/group-membership-routes.test.ts test/group-state/group-presence-routes.test.ts test/group-state/group-state-mutation-routes.test.ts test/group-state/group-state-openapi-contract.test.ts test/group-state/group-state-read-routes.test.ts test/group-state/group-state-route-errors.test.ts test/group-state/register-group-state-routes.test.ts test/routes/state-api-cross-feature-routes.test.ts test/db/postgres-queue-pubsub-bridge.test.ts test/db/in-memory-schema-bootstrap.test.ts test/db/pglite-sql-adapter.test.ts",
-    "test:deno": "cd apps/api-v1 && deno task test && cd ../rallar-black-box-control-server && deno task test && cd ../relic-hunter-server-v1 && deno task check && cd ../.. && deno test -A packages/tests/shared-test/scenario-black-box-rtc-config.test.ts packages/tests/shared-test/execute-black-box-rtc-client-provider.test.ts",
+    "test:deno": "cd apps/api-v1 && deno task test && cd ../rallar-black-box-control-server && deno task test && cd ../relic-hunter-server-v1 && deno task check && deno task test && cd ../.. && deno test -A packages/tests/shared-test/scenario-black-box-rtc-config.test.ts packages/tests/shared-test/execute-black-box-rtc-client-provider.test.ts",
     "test:e2e": "npm run test:rallar && npm run test:rallar:recipe-console",
     "test:full-stack": "npm run test:full-stack:memory",
     "test:full-stack:memory": "npm run test:rallar:full-stack:memory",
