+++ packages/relic-hunters/tsconfig.json
@@ -16,15 +16,6 @@
       "@relic-hunters/*": [
         "../../packages/relic-hunters/*"
       ],
-      "@shared-web/*": [
-        "../../packages/shared-web/*"
-      ],
-      "@shared-server/*": [
-        "../../packages/shared-server/*"
-      ],
-      "@shared-graph/*": [
-        "../../packages/shared-graph/*"
-      ],
       "@shared/*": [
         "../../packages/shared/*"
       ]
