+++ packages/shared-graph/tsconfig.json
@@ -13,9 +13,6 @@
     "allowImportingTsExtensions": true,
     "types": [],
     "paths": {
-      "@shared-web/*": [
-        "../../packages/shared-web/*"
-      ],
       "@shared-graph/*": [
         "../../packages/shared-graph/*"
       ],
