+++ packages/shared-test/black-box-runner/api-v1-black-box-run.mts
@@ -1,4 +1,3 @@
-/// <reference lib="deno.ns" />
 import {
   removeApiV1FairnessProofArtifact,
   verifyApiV1FairnessProof,
