+++ packages/shared-test/black-box-runner/managed-api/api-v1-managed-api-readiness.mts
@@ -1,4 +1,3 @@
-/// <reference lib="deno.ns" />
 import {
   BEARER_CREDENTIAL,
   MANAGED_SECRET_ENV_KEY,
