+++ packages/shared-test/tsconfig.json
@@ -15,17 +15,17 @@
       "node"
     ],
     "paths": {
-      "@shared-test/*": [
-        "../../packages/shared-test/*"
+      "@shared-graph/*": [
+        "../../packages/shared-graph/*"
       ],
       "@shared-server/*": [
         "../../packages/shared-server/*"
       ],
+      "@shared-test/*": [
+        "../../packages/shared-test/*"
+      ],
       "@shared-web/*": [
         "../../packages/shared-web/*"
-      ],
-      "@shared-graph/*": [
-        "../../packages/shared-graph/*"
       ],
       "@shared/*": [
         "../../packages/shared/*"
