+++ packages/shared-web/tsconfig.json
@@ -16,9 +16,6 @@
       "@shared-web/*": [
         "../../packages/shared-web/*"
       ],
-      "@shared-server/*": [
-        "../../packages/shared-server/*"
-      ],
       "@shared-graph/*": [
         "../../packages/shared-graph/*"
       ],
